requirements.txt
langchain
streamlit
sqlalchemy
openai
pandas
python-dotenv